Java Social Application
Code to create various classes that model features of social media. For example, a website like Facebook is rich in properties about its users that can be theoretically mined for beneficial purposes, such as predicting depression in someone based on their activity and offering support and help, or less benign, such as targeting advertisements. In particular, after successfully completing this project, following goals should be met:

An understanding of the instanceof keyword.
An understanding of casting.
An understanding of the limitations of the above two points.
An understanding of creating your own exception.
An understanding of catching and raising your own exception in Java.